fix(attach): respect env, envFile from IAttachRequestArgs
if `env` or `envFile` are specified as args for attach tasks they
should be respected. otherwise, the only way to set environment
variables when attaching is via a `.env` file. this is particularly
misleading because the documentation (and definition of
`IAttachRequestArgs`) imply that it should work.

currently the `runOptions` were never getting set when responding
to attach requests, resulting in the `env` and `envFile` check in
`Packager#start()` skipping the env update.

call `setRunOptions(attachArgs)` in `attachRequest` after
`appLauncher` and `appLauncher.packager` are initialized via the
call to `initializeSettings` to ensure `runOptions` is valid and
`env` or `envFile` can be correctly read.